+++ __init__.py
@@ -3,4 +3,4 @@ __author__ = 'Ian Housman'
 __email__ = 'ian.housman@gmail.com'
 
 #Version format yyyy.m.n
-__version__ = '2023.10.1'
+__version__ = '2023.12.1'