+++ examples/CCDCViz.py
@@ -62,8 +62,7 @@ Map.addLayer(ccdcImg,{'opacity':0},'Raw CCDC Output',False)
 
 #Extract the change years and magnitude
 changeObj = changeDetectionLib.ccdcChangeDetection(ccdcImg,changeDetectionBandName);
-changeDetectionLib.lossMagPalette = changeDetectionLib.lossMagPalette.split(',')
-changeDetectionLib.lossMagPalette.reverse()
+
 Map.addLayer(changeObj[sortingMethod]['loss']['year'],{'min':startYear,'max':endYear,'palette':changeDetectionLib.lossYearPalette},'Loss Year')
 Map.addLayer(changeObj[sortingMethod]['loss']['mag'],{'min':-0.5,'max':-0.1,'palette':changeDetectionLib.lossMagPalette},'Loss Mag',False);
 Map.addLayer(changeObj[sortingMethod]['gain']['year'],{'min':startYear,'max':endYear,'palette':changeDetectionLib.gainYearPalette},'Gain Year');
@@ -86,12 +85,15 @@ fittedBns = fitted.select(['.*_fitted']).first().bandNames()
 bns = fittedBns.map(lambda bn: ee.String(bn).split('_').get(0))
 
 # Filter down to the next to the last year and a summer date range
+compositeYear = endYear-1
 syntheticComposites = fitted.select(fittedBns,bns)\
-    .filter(ee.Filter.calendarRange(endYear-1,endYear-1,'year'))\
-    .filter(ee.Filter.calendarRange(190,250)).first()
+    .filter(ee.Filter.calendarRange(compositeYear,compositeYear,'year'))
+    # .filter(ee.Filter.calendarRange(190,250)).first()
 
 # Visualize output as you would a composite
-Map.addLayer(syntheticComposites,getImagesLib.vizParamsFalse,'Synthetic Composite')
+getImagesLib.vizParamsFalse['dateFormat']='YYMMdd'
+getImagesLib.vizParamsFalse['advanceInterval']='day'
+Map.addTimeLapse(syntheticComposites,getImagesLib.vizParamsFalse,f'Synthetic Composite Time Lapse {compositeYear}')
 ####################################################################################################
 #Load the study region
 studyArea = ccdcImg.geometry()