+++ examples/getCombinedLandsatSentinel2Wrapper.py
@@ -130,10 +130,10 @@ cloudHeights = ee.List.sequence(500,10000,500)
 #This method works really well
 applyCloudProbability = True
 
-#Whether to use the pre-computed cloudScore+ to mask
-#clouds and cloud shadows for Sentinel 2
-#This method works really well and should be used instead of all other methods once it finishes
-# running for the S2 archive (~ early 2024????)
+# Whether to use the pre-computed cloudScore+ to mask
+# clouds and cloud shadows for Sentinel 2
+# This method works really well and should be used instead of all other methods once it finishes
+# running for the S2 archive (~ spring 2024????)
 applyCloudScorePlusSentinel2 = False
 
 #If cloudProbability is chosen, choose a threshold